added some comments
